requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
alpha_vantage==3.0.0
asgiref==3.8.1
astroid==3.2.4
async-timeout==4.0.3
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.3.2
coverage==7.6.1
dill==0.3.8
dj-database-url==2.2.0
Django==5.1.1
django-cors-headers==4.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
exceptiongroup==1.2.2
frozenlist==1.4.1
idna==3.10
iniconfig==2.0.0
isort==5.13.2
mccabe==0.7.0
multidict==6.1.0
packaging==24.1
pillow==10.3.0
platformdirs==4.3.3
pluggy==1.5.0
polygon-api-client==1.14.2
psycopg==3.2.1
PyJWT==2.8.0
pylint==3.2.7
pylint-django==2.5.5
pylint-plugin-utils==0.8.2
pytest==8.3.3
pytest-cov==5.0.0
pytest-django==4.9.0
python-dotenv==1.0.1
requests==2.32.3
sqlparse==0.5.0
tomli==2.0.1
tomlkit==0.13.2
typing_extensions==4.12.2
urllib3==2.2.3
websockets==12.0
yarl==1.11.1